# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApproveBody": {
        "additionalProperties": false,
        "description": "``POST /trajectories/{id}/approve``'s optional body: a budget, for a ceiling raise only.",
        "properties": {
          "budget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RaiseBody"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ApproveBody",
        "type": "object"
      },
...
      "DenyBody": {
        "additionalProperties": false,
        "description": "``POST /trajectories/{id}/deny``'s optional body: the stated reason.",
        "properties": {
          "reason": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "DenyBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MoneyBody": {
        "additionalProperties": false,
        "description": "A money amount as ``{currency, nanos}``.",
        "properties": {
          "currency": {
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "nanos": {
            "minimum": 1.0,
            "title": "Nanos",
            "type": "integer"
          }
        },
        "required": [
          "nanos"
        ],
        "title": "MoneyBody",
        "type": "object"
      },
      "RaiseBody": {
        "additionalProperties": false,
        "description": "The new per-trajectory ceiling offered with a ``ceiling_raise`` grant.",
        "properties": {
          "money": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "tokens": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tokens"
          }
        },
        "title": "RaiseBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "PromptCadence",
    "version": "1.3.3"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/approvals": {
      "get": {
        "description": "Every pending request with its age, oldest first; ``status=all`` lists resolved ones too.\n\n``read`` scope. Each item names the trajectory, the kind of question (a held plan, a gated\nstep, the bypass gate, a scoped re-approval, a ceiling raise), the steps it is scoped to, what\nit asks (``detail``), when it expires, and how long it has waited.\n\n``status=all`` with no ``trajectory_id`` is every request ever raised, newest first, paged by\n``limit`` (clamped to 200) and ``cursor``; ``400 VALIDATION_ERROR`` for a cursor this API did\nnot mint. Every other combination answers as it always has, unpaged (row WPC1).",
        "operationId": "list_approvals_api_v1_approvals_get",
        "parameters": [
          {
            "in": "query",
            "name": "trajectory_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Trajectory Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pending approval requests",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/v1/trajectories/{trajectory_id}/approve": {
      "post": {
        "description": "T8: mint what the pending request asked for, under the caller's identity.\n\nRequires the ``approve`` scope (``401``/``403`` otherwise). Idempotent per request: granting\nan already-granted request returns ``200`` with ``already_resolved: true`` and changes\nnothing. ``409 APPROVAL_INVALID_STATE`` when nothing is pending or the last request was\ndenied or expired; ``400 VALIDATION_ERROR`` when a ceiling raise offers no budget or a budget\nis offered to a request that is not a raise.",
        "operationId": "post_approve_api_v1_trajectories__trajectory_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "trajectory_id",
            "required": true,
            "schema": {
              "title": "Trajectory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ApproveBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Grant the trajectory's pending approval request",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/v1/trajectories/{trajectory_id}/deny": {
      "post": {
        "description": "T9: halt the trajectory with the denial recorded, under the caller's identity.\n\nRequires the ``approve`` scope. Idempotent per request. ``409 APPROVAL_INVALID_STATE`` when\nnothing is pending and the last request was not denied.",
        "operationId": "post_deny_api_v1_trajectories__trajectory_id__deny_post",
        "parameters": [
          {
            "in": "path",
            "name": "trajectory_id",
            "required": true,
            "schema": {
              "title": "Trajectory Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DenyBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deny the trajectory's pending approval request",
        "tags": [
          "approvals"
        ]
      }
    },
...
  }
}